IFP: Only format the output name to the short version before output
The ifp_user_get_attr_done() request handler was reused for both
GetUserGroups and GetUserAttrs requests. Yet, it performed output
formatting of name and nameAlias.

This is bad, because the output formatting should really be done only
during output. Also, it broke any post-processing of the returned
message which the request might do later.

Reviewed-by: Pavel Březina <pbrezina@redhat.com>