TESTS: Integration test for idle timeout
The secrets responder test was chosen even though the bug was in the generic
responder code b/c it runs a single responder process, so it's trivial to
read the PID of the responder under test.

Changes subprocess.call() for os.execv() so that the setup function can
return the secret responder PID right away.

The client timeout in the test has to be at least 10 seconds because
internally, the responders don't allow a shorter timeout.

Regression test for https://pagure.io/SSSD/sssd/issue/3448

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>