PAM: handled multiple certs in the responder
This patch refactors the handling of the certificate and the attributes
to address the certificate on the Smartcard (module name, token name and
key id). Instead of using individual variables the values are put into a
new struct cert_auth_info. Since the new struct can be used as a list
the PAM responder can now handle multiple certificates on the Smartcard
and can send the needed data to pam_sss with multiple SSS_PAM_CERT_INFO
messages.

Unit tests are added to confirm the expected behavior.

Related to https://pagure.io/SSSD/sssd/issue/3560

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>
Tested-by: Scott Poore <spoore@redhat.com>
(cherry picked from commit 0bdd8800c16f39b8fe308d20694ad905c669dff3)

The following binaries have been removed from the original patch:
- src/tests/cmocka/p11_nssdb_2certs/cert9.db
- src/tests/cmocka/p11_nssdb_2certs/key4.db

The reason for that is that we can't apply a patch which is a binary
file using rpm, thus removing the patches here and adding them as source
files in the sssd.spec seems to be the best solution.