Fixes for backport of PR#57 to SSSD-1.13
SSSD-1.13 still uses different names in the cache for the configured
domain and sub-domains, hence some name related calls had to be
adjusted.

dp_target_enabled() is not available in 1.13 and I think just checking
the schema would be sufficient.

In more recent version of SSSD sysdb_group_membership_mod() treats the
domain argument only as the domain of the group and the domain of the
members are derived from the fully-qualified name. With this patch the
domain of the members is determined as well to allow to add members of
different domains to a group. This required a small addition to the unit
tests as well.