PAM: add certificate's label to the selection prompt
Some types of Smartcards contain multiple certificate with the same
subject-DN for different usages. To make it easier to choose between
them in case the matching rules allow more than one of them for
authentication the label assigned to the certificate on the Smartcard is
shown in the selection prompt as well.

Related to https://pagure.io/SSSD/sssd/issue/3560