ipa: handle users from different domains in ipa_resolve_user_list_send()
Instead of assuming that all users in the list can be found in the
provided domain with this patch the domain name part of the user name is
preferred. The provided domain name is used as a fallback.

Related to https://pagure.io/SSSD/sssd/issue/3579

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>
(cherry picked from commit 7988988aab5bd0249476671b850eb3909aa753f8)