overrides: fixes for sysdb_invalidate_overrides()
There were two issues in sysdb_invalidate_overrides().

First, SYSDB_CACHE_EXPIRE was only reset for the entry in the data cache
but not in the timestamp cache.

Second, if one of the steps in the combined replace and delete operation
failed no change was committed to the cache. If, for whatever reasons,
a user or group object didn't had SYSDB_OVERRIDE_DN set the delete
failed and hence SYSDB_CACHE_EXPIRE wasn't reset as well. To make sure
the cache is in a consistent state after a view change the replace and
the delete operations are don in two steps.

Related to https://pagure.io/SSSD/sssd/issue/3579

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>
(cherry picked from commit 4671acb949c65c5c080532e03b1b6f1c9377a6a5)