NSS: Don't generate encryption key if none is passed
Resolves: https://pagure.io/SSSD/sssd/issue/3614

The nss_ctx_init() function tried to be user-friendly and generate a
random key if none was passed. But the function that generated the
random key would access the opaque key object to get the raw key
material. And this fails with certain restricted environments e.g. when
the machine is running in a FIPS mode.