man: remove references to local provider
This commit is at least a little bit controversial as the way the xml
files are processed the include always happens and we cannot filter out
the content of the local.xml *after* its include.

In order to work this around we have to explicitly move the context of
the local.xml to each files where it's included and then use the
conditionals to remove it from the manuals.

As the local provider is already deprecated I do *not* expect
maintenance to be done os those duplicated parts apart from dropping all
of them altogether as soon as the dlocal provider is removed (which will
happen soon, hopefully).

Related:
https://pagure.io/SSSD/sssd/issue/3304

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>