AD: consider resource_groups in PAC as well
With recent versions of Active Directory the SIDs of Domain Local groups
might be only available in the resource_groups section of the PAC, this
feature is also called SID compression. To get a complete list of groups
the user is a member of the SIDs from this section must be extracted as
well.

Resolves https://pagure.io/SSSD/sssd/issue/3767

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
(cherry picked from commit 13c8450788a429fa49ba532b40ebfd7f3a4132e4)

DOWNSTREAM:
Resolves: rhbz#1592964 - Groups go missing with PAC enabled in sssd