tests: adapt sysdb-tests for disabling local provider
local provider has mpg enabled by default but, as the local provider is
only conditionally built nowadays (and disabled by default) we have to
set mpg to true when running some tests and the local provider is not
built.

Also, we have to adapt functions that would be used to test and/or
internally take advantage of the automatically generated uid/gid for the
user as it's specific for local provider.

Related:
https://pagure.io/SSSD/sssd/issue/3304

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>